+++ src/JSEngine.hpp
@@ -46,7 +46,7 @@ namespace be {
         static JSValue getGlobalValue(const char * name) ;
 
         JSValue eval(const char * code, int code_len=-1, const char * filename="eval", int flags=JS_EVAL_TYPE_GLOBAL) ;
-        JSValue evalScript(const char * filepath, int flags=JS_EVAL_TYPE_MODULE) ;
+        JSValue evalScript(const char * filepath, int flags=JS_EVAL_TYPE_MODULE, bool dumpException=true) ;
         void runScript(const char * filepath, int flags=JS_EVAL_TYPE_MODULE) ;
 
         static JSEngine * fromJSContext(JSContext *) ;